Switching projectiles works!

+++ src/main/java/ca/qc/bdeb/sim203/projetjavafx/Main.java
@@ -7,9 +7,9 @@ public class Main extends Application {
     public static void main(String[] args) {
         launch(args);
     }
-    public static final int LARGEUR_ECRAN = 900;
-    public static final int LARGEUR_MONDE = 8*Main.LARGEUR_ECRAN;
-    public static final int HAUTEUR = 520;
+    public static final double LARGEUR_ECRAN = 900;
+    public static final double LARGEUR_MONDE = 8*Main.LARGEUR_ECRAN;
+    public static final double HAUTEUR = 520;
     private Stage stage;
     @Override
     public void start(Stage primaryStage) throws Exception {
